pyproject.toml
[project]
name = "company-insight-agent"
version = "0.1.0"
description = "An AI agent designed to research a company"
authors = [{ name = "Mike Charles", email = "serverlessmike@gmail.com" }]
requires-python = ">=3.12"

# Runtime dependencies here
dependencies = [
    "ipykernel",
    "openai-agents",
    "gradio",
    "python-dotenv",
    "pydantic",
]
